# docs/swagger.yaml
definitions:
  control-panel-service_internal_server_dto_response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  entity.MotherServiceStatus:
    enum:
    - pending
    - running
    - paused
    - stopped
    - deleted
    type: string
    x-enum-comments:
      MotherServiceStatusDeleted: stop and delete containers. can not start again.
      MotherServiceStatusPaused: application level pause on sending request
      MotherServiceStatusPending: mother service just created
      MotherServiceStatusRunning: test is running on application level (sending level)
      MotherServiceStatusStopped: stop container but can start scenario again.
    x-enum-descriptions:
    - mother service just created
    - test is running on application level (sending level)
    - application level pause on sending request
    - stop container but can start scenario again.
    - stop and delete containers. can not start again.
    x-enum-varnames:
    - MotherServiceStatusPending
    - MotherServiceStatusRunning
    - MotherServiceStatusPaused
    - MotherServiceStatusStopped
    - MotherServiceStatusDeleted
...
  request.MotherService:
    properties:
      database_name:
        type: string
      database_table_name:
        type: string
      exception_rate:
        type: integer
      name:
        type: string
      random_response_delay_max:
        type: integer
      random_response_delay_min:
        type: integer
      response_delay_duration:
        type: integer
      response_delay_rate:
        type: integer
    type: object
  request.PaginationRequest:
    properties:
      page:
        type: integer
      per_page:
        type: integer
    type: object
...
  response.MotherService:
    properties:
      created_at:
        type: string
      database_name:
        type: string
      database_table_name:
        type: string
      exception_rate:
        type: integer
      id:
        type: integer
      name:
        type: string
      random_response_delay_max:
        type: integer
      random_response_delay_min:
        type: integer
      response_delay_duration:
        type: integer
      response_delay_rate:
        type: integer
      service_deployment_address:
        type: string
      status:
        $ref: '#/definitions/entity.MotherServiceStatus'
      updated_at:
        type: string
    type: object
  response.MotherServiceResponseByID:
    properties:
      data:
        $ref: '#/definitions/response.MotherService'
    type: object
  response.PaginatedMotherServices:
    properties:
      data:
        items:
          $ref: '#/definitions/response.MotherService'
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
  response.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/mother-services:
    post:
      consumes:
      - application/json
      description: Create a new mother service
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.MotherService'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
      summary: Create a mother service
      tags:
      - mother-services
  /api/v1/mother-services/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific mother service by its ID
      parameters:
      - description: Mother service ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MotherServiceResponseByID'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
      summary: Get mother service by ID
      tags:
      - mother-services
  /api/v1/mother-services/{id}/delete:
    post:
      consumes:
      - application/json
      description: Delete a specific mother service by its ID.
      parameters:
      - description: Test scenario ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
      summary: Delete a mother service.
      tags:
      - mother-services
  /api/v1/mother-services/search:
    post:
      consumes:
      - application/json
      description: Get mother services with pagination support
      parameters:
      - description: Pagination request with page and per_page
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.PaginationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginatedMotherServices'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/control-panel-service_internal_server_dto_response.ErrorResponse'
      summary: Get paginated mother services
      tags:
      - mother-services
...
swagger: "2.0"